Move plugin source parsing into plugins service
Move plugin source parsing and release version validation out of gestaltd/internal/pluginsource and into services/plugins/source.

Interface diff:

- import "github.com/valon-technologies/gestalt/server/internal/pluginsource"

- pluginsource.Parse(raw)

- pluginsource.ValidateVersion(version)

+ import "github.com/valon-technologies/gestalt/server/services/plugins/source"

+ source.Parse(raw)

+ source.ValidateVersion(version)

Example:

src, err := source.Parse(manifest.Source)

if err := source.ValidateVersion(version); err != nil { return err }
