+++ gestaltd/services/plugins/source/source_test.go
@@ -1,4 +1,4 @@
-package pluginsource
+package source
 
 import (
 	"testing"
