+++ gestaltd/services/plugins/source/version.go
@@ -1,4 +1,4 @@
-package pluginsource
+package source
 
 import (
 	"fmt"
@@ -10,14 +10,14 @@ import (
 func ValidateVersion(version string) error {
 	canonical := versionPrefix + version
 	if !semver.IsValid(canonical) {
-		return fmt.Errorf("pluginsource: invalid semver %q", version)
+		return fmt.Errorf("plugin source: invalid semver %q", version)
 	}
 	base := canonical
 	if i := strings.IndexByte(base, '+'); i != -1 {
 		base = base[:i]
 	}
 	if semver.Canonical(canonical) != base {
-		return fmt.Errorf("pluginsource: invalid semver %q", version)
+		return fmt.Errorf("plugin source: invalid semver %q", version)
 	}
 	return nil
 }
